requirements.txt
ipykernel==6.25.2
notebook==7.0.3
numpy==1.25.2
opencv-python==4.8.0.76
Pillow==10.0.0
pytest==7.4.2
tqdm==4.66.1
